requirements.txt
a2a-types==0.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
deprecated==1.2.18
distro==1.9.0
fast-agent-mcp==0.2.19
fastapi==0.115.12
frozenlist==1.6.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.6.1
jiter==0.9.0
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
multidict==6.4.3
openai==1.76.0
opentelemetry-api==1.32.1
opentelemetry-distro==0.53b1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-anthropic==0.39.3
opentelemetry-instrumentation-openai==0.39.3
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-semantic-conventions-ai==0.4.3
packaging==25.0
prompt-toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pygments==2.19.1
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
